requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
idna==3.10
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pantab==5.2.0
pip==24.2
pyarrow==18.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
tableauserverclient==0.34
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
